requirements.txt
scapy==2.3.1
<<<<<<< HEAD
pypcapfile==0.10
dpkt==1.8.6.2
requests==2.7.0
=======
requests==2.7.0
>>>>>>> HTTPUtils
